Removed ExceptionWrapper from serializers. This is done since it is improper to throw a parse exception from serializers. Moreover, the serializers dont deal with external input, they deal with classes, so they need not guard against spurious input.

+++ trunk/gson/src/test/java/com/google/gson/functional/MapTest.java
@@ -95,8 +95,8 @@ public class MapTest extends TestCase {
     }).create();
     String json = "{\"a\":1,\"b\":2}";
     MyParameterizedMap<String, Integer> map = gson.fromJson(json, type);
-    assertEquals(1, ((Integer) map.get("a")).intValue()); 
-    assertEquals(2, ((Integer) map.get("b")).intValue()); 
+    assertEquals(1, map.get("a").intValue()); 
+    assertEquals(2, map.get("b").intValue()); 
   }
 
   private static class MyParameterizedMap<K, V> extends LinkedHashMap<K, V> {
